Update SymbolTable and integrate it into Scanner
SymbolTable has additional functionality to keep track of multiple
symbol tables, and Scanner adds functions and identifiers to the correct
symbol table based on the scope of the current position.

If the current line position is within a function, the scanner will add
identifiers to that new symbol table. If the current position is global,
it will add identifiers to the global symbol table or any outer symbol
table based on the depth of functions.

Scanner is now aware of arrays, which may need to move to the
parser depending on how complex the declarations are.