Fix else cond executing when if cond is true
bExec was not being used to control the else condition. The true/false
values were hard coded, so it was always executing.

Some other condition control is hard coded in the ifStmt function, so
that may need to be changed to use bExec as well.

All of the output for Program 3 is correct now.